+++ src/main/java/com/strangeone101/pixeltweaks/TweaksConfig.java
@@ -14,6 +14,9 @@ public class TweaksConfig {
     public static ForgeConfigSpec.ConfigValue<Boolean> randomlyDropRipeApricorns;
     public static ForgeConfigSpec.ConfigValue<Boolean> foxesLoveBerries;
     public static ForgeConfigSpec.ConfigValue<Boolean> antiTrampleCrops;
+    public static ForgeConfigSpec.ConfigValue<Integer> zygardeSpawnTime;
+    public static ForgeConfigSpec.ConfigValue<Integer> zygardeLuckyAttemptRate;
+    public static ForgeConfigSpec.ConfigValue<Integer> zygardeStormCoreRate;
     public static ForgeConfigSpec.ConfigValue<Double> catchMultiplier;
     public static ForgeConfigSpec.ConfigValue<Double> legendaryCatchMultiplier;
     public static ForgeConfigSpec.ConfigValue<Boolean> includeMythicals;
@@ -51,6 +54,15 @@ public class TweaksConfig {
 
         antiTrampleCrops = BUILDER.comment("If pokemon should be prevented from trampling crops. Default: true")
                 .define("antiTrampleCrops", true);
+
+        zygardeSpawnTime = BUILDER.comment("How often a Zygarde cell should attempt to spawn (in seconds). Does not guarantee a spawn. Set to 0 to disable. Default: 30")
+                .defineInRange("zygardeSpawnTime", 30, 0, 9999);
+
+        zygardeLuckyAttemptRate = BUILDER.comment("The chance of a spawn attempt being 'lucky'. Lucky attempts will spawn with a higher chance of being near the player. Chance is 1/<rate> Set to 0 to disable. Default: 4.")
+                .defineInRange("zygardeLuckyAttemptRate", 4, 0, 9999);
+
+        zygardeStormCoreRate = BUILDER.comment("The rate in which Zygarde Cores should spawn when it's stormy. Rate is 1/<rate>. Set to 0 to disable. Default: 7")
+                .defineInRange("zygardeStormCoreRate", 7, 0, 9999);
 
         BUILDER.pop();
         /*BUILDER.comment("All catch related features").push("Server Features - Catching");